Added: -Skill level button, enhanced skill level management -Skills take new parameter: their level -skills damages formulas, only need to code it in c -new sound : need mana -zombie spawning now varies depending on your level maybe more but don't remember

+++ diablo_project/source/ulScreenDisplay.c
@@ -32,7 +32,7 @@ void myulInitData (bool again)
 
 
 
-int myulCreateSprite (u8 data,int x,int y, s8 prio)
+int myulCreateSprite (u8 data,int x,int y, int prio)
 {
     int i=0, nb=-1;
     for (i=0; i<MAX_SPRITES ; i++)
